Finish Tips & BottomBar

+++ lib/models/spaceModel.dart
@@ -2,7 +2,8 @@ class Space {
   int id;
   String name;
   String imageURL;
-  double price;
+  // double price;
+  int price;
   String city;
   String country;
   int rating;
